Restructure things a bit and add support for log4v2 xml files
